fix
-pagar cuenta.

+++ database/migrations/tenant/2025_07_01_102554_create_payment_method_account_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('payment_method_accounts', function (Blueprint $table) {
-            $table->unsignedInteger('payment_method_id');
+            $table->unsignedBigInteger('payment_method_id');
             $table->unsignedBigInteger('bank_account_id');
             $table->timestamps();
